rename folder 'training' → 'GP_model'
